Shift mock tests to their own file
